day 9

+++ src/main/scala/aoc/day2.scala
@@ -9,8 +9,8 @@ import Option.when
 object Day2 with
 
   def initialise(noun: Int, verb: Int) =
-    RIO.accessM((tape: IArray[Int]) =>
-      Task.effect(tape.updated(1, noun).updated(2, verb)))
+    RIO.accessM((tape: IArray[Long]) =>
+      Task.effect(tape.updated(1, noun.toLong).updated(2, verb.toLong)))
 
   def find(n: Int, v: Int)(goal: Int) =
     run(n,v).fold(none, x => when(x == goal)(f"$n%2d$v%2d"))
